requirements.txt
easyocr==1.7.1
matplotlib==3.7.2
numpy==1.23.5
pandas==2.0.3
Pillow==10.0.1
Pillow==10.1.0
scikit_learn==1.3.0
streamlit==1.29.0
